+++ _templates/component/new/component.ejs.t
@@ -4,7 +4,7 @@ to: components/<%= name %>/<%= name %>.js
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import css from './<%= name %>.scss';
+import css from './<%= name %>.module.scss';
 
 const <%= name %> = ({ className }) => {
   return <div className={[css.<%= h.changeCase.camel(name) %>, className || ''].join(' ')}></div>;