+++ components/Header/Header.js
@@ -2,8 +2,10 @@ import { Component } from 'react';
 import PropTypes from 'prop-types';
 
 import Menu from '../Menu';
-import Link from '../Link';
-// import './Header.css';
+// import Link from '../Link';
+import DXLabLogo from '../DXLabLogo';
+import SLNSWLogo from '../SLNSWLogo';
+import MenuIconButton from '../MenuIconButton/MenuIconButton';
 
 // const menuItems = [
 //   { name: 'Home', url: '/' },
@@ -49,13 +51,7 @@ class Header extends Component {
 
     return (
       <header className="header">
-        <div className="logo logo--dxlab">
-          <Link href="/">
-            <a href="/">
-              <img src="/images/logo-dxlab.png" alt="DX Lab Logo." />
-            </a>
-          </Link>
-        </div>
+        <DXLabLogo className="logo logo--dxlab" />
 
         <nav
           className={`header__nav ${this.state.isMenuOpen ? 'is-active' : ''}`}
@@ -76,38 +72,17 @@ class Header extends Component {
           />
         </nav>
 
-        <div className="logo logo--slnsw">
-          <a
-            href="http://sl.nsw.gov.au"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <img
-              // className="logo logo--slnsw"
-              src="/images/logo-slnsw-white.png"
-              alt="State Library Logo."
-            />
-          </a>
-        </div>
+        <SLNSWLogo className="logo logo--slnsw" />
 
-        <button
+        <MenuIconButton
+          isOpen={this.state.isMenuOpen}
           id="primary-menu-button"
-          className={`primary-menu-button ${
-            this.state.isMenuOpen ? 'is-open' : ''
-          }`}
           aria-haspopup="true"
           aria-controls="primary-menu"
-          aria-expanded="false"
+          aria-expanded={this.state.isMenuOpen}
           aria-label="primary menu"
           onClick={this.handleMenuToggle}
-        >
-          <div>
-            <span />
-            <span />
-            <span />
-            <span />
-          </div>
-        </button>
+        />
       </header>
     );
   }